added typings for animate
