+++ tests/unit/mixins/Animatable.ts
@@ -2,7 +2,8 @@ import global from '@dojo/shim/global';
 import { VNode } from '@dojo/interfaces/vdom';
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
-import { AnimatableMixin, AnimationPlayer, AnimationControls, AnimationTimingProperties } from '../../../src/mixins/Animatable';
+import { AnimatableMixin, AnimationPlayer } from '../../../src/mixins/Animatable';
+import { AnimationControls, AnimationTimingProperties } from '../../../src/interfaces';
 import { WidgetBase } from '../../../src/WidgetBase';
 import { v } from '../../../src/d';
 import { spy, stub } from 'sinon';
